Added support for pyxl's .append method to support generators as arguments
Summary: You can now do things like <frag>{foo(x) for x in bar}</frag> without the noise of needing brackets for the list. The append method really only uses the iterable interface anyway, so if this was Python 2.6, it should have been testing if its input was a subclass of collections.Iterable rather than explicitly checking if its input was a tuple or list. But since this is Python 2.5, we can live with this :-)

Reviewed By: tido

CC: Server-Reviews