Add an invertible codec variant (#12)
The normal pyxl codec is irritatingly incompatible with syntax-tree based code
modification tools, since even if they read in their input using the pyxl codec,
they lose it in the output.

In order to allow using tools like pyannotate, modernize, and black, we add an
invertible pyxl codec that produces python code that can be reversed back into pyxl.

Basically, ignoring many details, we turn
```
<foo>
    <bar baz="{baz}" spam="{spam}" eggs="{eggs}" />
</foo>
```
into
```
html.PYXL('''<foo>
    <bar baz="{{{}}}" spam="{{{}}}" eggs="{{{}}}" />
</foo>''', baz , spam , eggs )
```
from which we can reconstruct the original, even if the arguments get modified.
Additionally, we actually store more information to support reindenting the html if its start column changes due to black or some similar tool, and the actual generated pyxl code is also in there to allow using tools that remove unused imports.

Some substantial refactors are unfortunately lumped in here.

This adds tests that every file in the tests directory

 * Round-trips via the new pyxl codec without any changes
 * Produces the same AST after being formatted with black
 * Is idempotent when formatting with black

I also ran these tests against all of the pyxl code we have internally.
There is one bug revealed by this that is not fixed, which is that we will
spuriously insert extra space into multi-line string literals in a python
fragment. We only do that a single digit number of times so I'm going
to call it good enough.